requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

